Finished refaktoring
