+++ backend/requirements.txt
@@ -10,7 +10,5 @@ httpx
 qrcode
 fpdf
 Pillow>=9.0
-escpos
-pdf2image
 python-socketio[asyncio_client]
 asyncio
