+++ backend/src/PDF_creator/PresidentTicket.py
@@ -1,9 +1,9 @@
-from fpdf import FPDF
+import os
 import qrcode
-from PIL import Image
+
+from fpdf import FPDF
 
 from src.PDF_creator.BaseTicket import BaseTicket
-
 
 class PresidentTicket(BaseTicket):
     """ Class for ticket from vote with President candidate """
