+++ docker-compose.yml
@@ -10,13 +10,14 @@ services:
     restart: unless-stopped
     environment:
       - ROOT_PATH=/backend
-      - VOTING_SERVICE_PATH=host.docker.internal:8080/voting-service-api
-      - STATE_VECTOR_PATH=host.docker.internal:8080/statevector
-      - VOTING_PROCESS_MANAGER_HOST=host.docker.internal:8080
-      - VOTING_PROCESS_MANAGER_HOST_SOCKET_PATH=voting-process-manager-api/ws/socket.io
-      - VOTING_PROCESS_MANAGER_PATH=host.docker.internal:8080/voting-process-manager-api
+      - HYPERTEXT_PROTOCOL=https://
+      - VOTING_SERVICE_PATH=team17-21.studenti.fiit.stuba.sk/gateway/voting-service-api
+      - STATE_VECTOR_PATH=team17-21.studenti.fiit.stuba.sk/gateway/statevector
+      - VOTING_PROCESS_MANAGER_HOST=team17-21.studenti.fiit.stuba.sk
+      - VOTING_PROCESS_MANAGER_HOST_SOCKET_PATH=gateway/voting-process-manager-api/ws/socket.io
+      - VOTING_PROCESS_MANAGER_PATH=team17-21.studenti.fiit.stuba.sk/gateway/voting-process-manager-api
       - PRINTER_IP_ADDRESS=192.168.0.168
-      - VT_ONLY_DEV=1
+      - VT_ONLY_DEV=0
       - DONT_WAIT_FOR_TOKEN=1
 
   token-reader:
